Move test files and resources into test folder
