+++ packages/sdc-populate/src/test/ReadInitialExpressions.test.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { readPopulationExpressions } from './ReadPopulationExpressions';
+import { readPopulationExpressions } from '../ReadPopulationExpressions';
 import type { Questionnaire } from 'fhir/r4';
 import Q715XFhirQuery from './resources/715-v.json';
